+++ src/components/map/MapSigmets.vue
@@ -212,7 +212,7 @@ const styles = {
 function handleMapClick(event: MapBrowserEvent<any>) {
     openSigmet.value = null;
     const features = map.value?.getFeaturesAtPixel(event.pixel, { hitTolerance: 2 });
-    if (!features?.every(x => x instanceof RenderFeature || x.getProperties().dataType || x.getProperties().type === 'local')) return;
+    if (!features?.every(x => x instanceof RenderFeature || x.getProperties().dataType || x.getProperties().type === 'local' || x.getProperties().type === 'root')) return;
 
     const sigmets = features.filter(x => x.getProperties()?.dataType);
     if (!sigmets.length) return;