+++ src/server/api/data/debug/[type]/[id]/compare.ts
@@ -1,5 +1,6 @@
 import { handleH3Error } from '~/utils/backend/h3';
-import { getDiffPolygons, getSimAwareData } from '~/utils/backend/debug/data-get';
+import { getDiffPolygons, getSimAwareData, getVatSpyData } from '~/utils/backend/debug/data-get';
+import { vatspyDataToGeojson } from '~/utils/backend/vatsim/vatspy';
 
 export default defineEventHandler(async event => {
     const { type, id } = getRouterParams(event);
@@ -8,6 +9,12 @@ export default defineEventHandler(async event => {
             event,
             statusCode: 400,
         });
+    }
+
+    if (type === 'vatspy') {
+        const data = await getVatSpyData(+id);
+
+        return getDiffPolygons(vatspyDataToGeojson(data), 'vatspy');
     }
 
     return getDiffPolygons(await getSimAwareData(+id), 'simaware');